packages/cockpitdecks_xp/pyproject.toml
# ###########################################
#
# Build
#
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "/cockpitdecks_xp"
]

[tool.hatch.version]
path = "cockpitdecks_xp/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true


# ###########################################
#
# Project
#
[project]
name = "cockpitdecks_xp"

authors = [
  { name="Pierre M.", email="pierre@devleaks.be" },
  { name="Duane Licudi", email="dlicudi@gmail.com" }
]

description = "Cockpitdecks X-Plane Simulator Interface"

readme = "README.md"

license = {file = "LICENCE"}

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Simulation",
]

requires-python = ">=3.10"

dependencies = [
    "cockpitdecks",
    "xpwebapi",
    "metar~=1.11.0",
    "avwx-engine[scipy]~=1.9.0",
]

dynamic = [
    "version"
]

[project.optional-dependencies]
development = [
    "tabulate~=0.9.0"
]

[project.urls]
Homepage = "https://github.com/dlicudi/cockpitdecks"
Documentation = "https://devleaks.github.io/cockpitdecks-docs/"
Issues = "https://github.com/dlicudi/cockpitdecks/issues"
Source = "https://github.com/dlicudi/cockpitdecks"

[tool.uv.sources]
cockpitdecks = { workspace = true }
xpwebapi = { workspace = true }

[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
]
ignore = []

[tool.black]
line-length = 160

[tool.flake8]
max-line-length = 160

